refactor(server): split worker runner responsibilities
