+++ apps/server/src/execution/worker-runner-types.ts
@@ -1,0 +1,42 @@
+import type { ChatAgent } from "@openomni/agent";
+import type { Auth } from "@openomni/llm";
+import type { BackgroundManager, InjectionQueue } from "@openomni/openomni";
+import type { WorkerBootstrap } from "@openomni/protocol";
+import type { WorkerRunState } from "./worker-run-state";
+import type { WorkerRunIpcServer } from "./worker-runner-ipc";
+
+type WorkerRunnerChatAgentOptions = Parameters<typeof ChatAgent.create>[0];
+type WorkerRunnerAgent = Pick<ReturnType<typeof ChatAgent.create>, "run">;
+
+interface WorkerRunnerEnvironment {
+  readonly ipcAuthToken: string;
+  readonly workerId: string;
+  readonly server: WorkerRunIpcServer;
+  readonly activeRuns: WorkerRunState.ActiveRunRegistry;
+  readonly bootstrapReady: Promise<void>;
+  readonly backgroundManager: ReturnType<typeof BackgroundManager.create>;
+  readonly injectionQueue: InjectionQueue.Instance;
+  readonly defaultWorkspaceRoot: string | undefined;
+  readonly getBootstrap: () => WorkerBootstrap.Bootstrap | null;
+  readonly resolveAuth: (provider: string) => Auth.Info | undefined;
+  readonly createAgent?: (options: WorkerRunnerChatAgentOptions) => WorkerRunnerAgent;
+}
+
+export interface WorkerRunnerSpawnOptions extends WorkerRunnerEnvironment {
+  readonly params: Record<string, unknown> | undefined;
+  readonly respond: (result: unknown) => void;
+}
+
+export function respondSpawnRejected(options: {
+  readonly params: Record<string, unknown> | undefined;
+  readonly respond: (result: unknown) => void;
+  readonly error: string;
+}): void {
+  const { params, respond, error } = options;
+  respond({
+    runId: typeof params?.runId === "string" ? params.runId : "unknown",
+    sessionId: typeof params?.sessionId === "string" ? params.sessionId : "unknown",
+    status: "failed",
+    error,
+  });
+}
